% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_wskazniki_admin_1.R
\name{agreguj_bs1_admin_1}
\alias{agreguj_bs1_admin_1}
\title{Obliczanie wskaznikow na poziomie zagregowanym dla 1. fali monitoringu
\itemize{
\item dane administracyjne
}}
\usage{
agreguj_bs1_admin_1(wsk2, wsk3, wsk4, grupy, raport, monitoring = 2021)
}
\arguments{
\item{wsk2}{ramka danych z tabeli pośredniej nr 2 (P2) z wynikami z 1.
rundy monitoringu na danych administracyjnych}

\item{wsk3}{ramka danych z tabeli pośredniej nr 3 (P3) z wynikami z 1.
rundy monitoringu na danych administracyjnych}

\item{wsk4}{ramka danych z tabeli pośredniej nr 4 (P4) z wynikami z 1.
rundy monitoringu na danych administracyjnych}

\item{grupy}{ramka danych zawierająca definicje podziałów na grupy -
np. zwrócona przez funkcję \code{\link{utworz_grupowanie_ze_zmiennej}}}

\item{raport}{rodzaj raportu: po roku (wartość 0) lub po 2 latach (wartość 1)}

\item{monitoring}{rok, w którym przeprowadzany jest monitoring}
}
\value{
data frame
}
\description{
Tylko dla Branżowych Szkół 1. stopnia
}
\seealso{
\code{\link{agreguj_wskazniki}} oraz przekazywane do niej funkcje
używane do obliczania konkretnych wskaźników zagregowanych:
\itemize{
\item{\code{\link{dane_szkoly_ad1}},}
\item{\code{\link{l_abs_ad1}},}
\item{\code{\link{l_kobiet_ad1}},}
\item{\code{\link{l_abs_zrodla_ad1}},}
\item{\code{\link{formy_ad1}},}
\item{\code{\link{S3_mies_ad1}},}
\item{\code{\link{S2_mies_ad1}},}
\item{\code{\link{zawody_S3_ad1}},}
\item{\code{\link{E2_nauka_kontyn_ad1}},}
\item{\code{\link{liczebnosc_branze_bs1_ad1}},}
\item{\code{\link{Z4_ods_prac_mies_ad1}},}
\item{\code{\link{Z8_formy_prac_mies_ad1}},}
\item{\code{\link{Z9_kont_mlod_ad1}},}
\item{\code{\link{W1_sr_doch_ad1}},}
\item{\code{\link{W3_sr_doch_uop_ad1}},}
\item{\code{\link{liczebnosc_branze_ucz_bs1_ad1}}}
}
}
