% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_admin_1.R
\name{B2_ods_bezrob_ad1}
\alias{B2_ods_bezrob_ad1}
\title{Obliczanie wskaznikow dla 1. fali monitoringu - dane administracyjne}
\usage{
B2_ods_bezrob_ad1(x, raport = 0, od = 9, do = 12)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{raport}{rodzaj raportu: po roku (wartość 0) lub po 2 latach (wartość 1)}

\item{od}{początek okresu wyliczania wskaźnika wyrażony miesiącem}

\item{do}{koniec okresu wyliczania wskaźnika wyrażony miesiącem}
}
\value{
lista
}
\description{
Funkcja licząca na potrzeby szablonu raportu rozkład liczby
miesięcy bezrobocia rejestrowanego wśród absolwentów w danym okresie. W
raporcie w 2021 roku będzie to "Liczba miesięcy bezrobocia rejestrowanego od
\emph{września 2020} do \emph{grudnia 2020} roku.
}
