% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_admin_1.R
\name{dane_szkoly_ad1}
\alias{dane_szkoly_ad1}
\title{Obliczanie wskaznikow dla 1. fali monitoringu - dane administracyjne}
\usage{
dane_szkoly_ad1(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
tekst
}
\description{
Funkcja przechowująca nazwę szkoły.
}
