% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_admin_1.R
\name{formy_ad1}
\alias{formy_ad1}
\title{Obliczanie wskaznikow dla 1. fali monitoringu - dane administracyjne}
\usage{
formy_ad1(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
lista
}
\description{
Funkcja tworząca i przechowująca listę wyrażeń tekstowych
używanych do podstawiania w szablonach raportów:
\itemize{
\item{Nazwa województwa w dopełniaczu przechowywana jako tekst}
}
}
