% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_admin_1.R
\name{l_abs_ad1}
\alias{l_abs_ad1}
\title{Obliczanie wskaznikow dla 1. fali monitoringu - dane administracyjne}
\usage{
l_abs_ad1(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
liczba
}
\description{
Funkcja oblicza wskaźnik opisujący liczbę absolwentów na danym
poziomie agregacji, którzy zostali objęci monitoringiem.
}
