% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_admin_1.R
\name{licz_zawody_ad1}
\alias{licz_zawody_ad1}
\title{Obliczanie wskaznikow dla 1. fali monitoringu - dane administracyjne}
\usage{
licz_zawody_ad1(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
lista
}
\description{
Funkcja licząca rozkład liczebności w zawodach na potrzeby
raportu wojewódzko-branżowego.
}
