% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_admin_1.R
\name{liczebnosc_branze_bs1_ad1}
\alias{liczebnosc_branze_bs1_ad1}
\title{Obliczanie wskaznikow dla 1. fali monitoringu - dane administracyjne}
\usage{
liczebnosc_branze_bs1_ad1(x, branza_kont_df, raport = 0, mies = 12)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{branza_kont_df}{ramka danych zawierająca informację o kontynuowaniu
kształcenia w danej branży (tabela danych pośrednich P2 lub zawierająca
analogiczne informacje oraz te same nazwy kolumn co tabela P2)}

\item{raport}{rodzaj raportu: po roku (wartość 0) lub po 2 latach (wartość 1)}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik - domyślnie
grudzień}

\item{kont_bs2_df}{ramka danych zawierająca informację o kontynuowaniu
kształcenia w szkołach branżowych 2. stopnia (tabela danych pośrednich P3 lub
zawierająca analogiczne informacje oraz te same nazwy kolumn co tabela P3)}
}
\value{
lista

lista
}
\description{
Funkcja licząca rozkład liczebności absolwentów kontynuujących
naukę w szkołach branżowych 2. stopnia w podziale na branże. Dodatkowo,
funkcja liczy liczebności absolwentów w branżach, a informacja ta służy jako
podstawa do definiowania warunków w szablonie raportu dla szkół branżowych 1.
stopnia. Funkcja ma sens tylko dla absolwentów szkół branżowych 1. stopnia,
ponieważ tylko oni mogą kontynuować naukę w szkołach branżowych 2. stopnia.

Funkcja licząca rozkład liczebności absolwentów kontynuujących
naukę w szkołach branżowych 2. stopnia w podziale na branże. Dodatkowo,
funkcja liczy liczebności absolwentów w branżach, a informacja ta służy jako
podstawa do definiowania warunków w szablonie raportu dla szkół branżowych 1.
stopnia. Funkcja ma sens tylko dla absolwentów szkół branżowych 1. stopnia,
ponieważ tylko oni mogą kontynuować naukę w szkołach branżowych 2. stopnia.
}
