% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_admin_1.R
\name{liczebnosc_branze_ucz_bs1_ad1}
\alias{liczebnosc_branze_ucz_bs1_ad1}
\title{Obliczanie wskaznikow dla 1. fali monitoringu - dane administracyjne}
\usage{
liczebnosc_branze_ucz_bs1_ad1(x)
}
\arguments{
\item{x}{ramka danych pośrednich P4}
}
\value{
lista
}
\description{
Funkcja licząca rozkład liczebności wyuczonych zawodów wśród
absolwentów w podziale na branże. Dodatkowo, funkcja liczy liczebności
absolwentów w branżach, a informacja ta służy jako podstawa do definiowania
warunków w szablonie raportu dla szkół branżowych 1. stopnia.
}
