% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_admin_1.R
\name{liczebnosc_dziedziny_tech_ad1}
\alias{liczebnosc_dziedziny_tech_ad1}
\title{Obliczanie wskaznikow dla 1. fali monitoringu - dane administracyjne}
\usage{
liczebnosc_dziedziny_tech_ad1(x, dziedzina_kont_df, raport = 0, mies = 12)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{dziedzina_kont_df}{ramka danych zawierająca informację o kontynuowaniu
kształcenia w danej dziedzinie (tabela danych pośrednich P2 lub zawierająca
analogiczne informacje oraz te same nazwy kolumn co tabela P2)}

\item{raport}{rodzaj raportu: po roku (wartość 0) lub po 2 latach (wartość 1)}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik - domyślnie
grudzień}
}
\value{
lista
}
\description{
Funkcja licząca rozkład liczebności absolwentów kontynuujących
naukę na studiach w podziale na dziedziny. Funkcja liczy wskaźnik tylko dla
absolwentów techników.
}
