% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_admin_1.R
\name{S2_mies_ad1}
\alias{S2_mies_ad1}
\title{Obliczanie wskaznikow dla 1. fali monitoringu - dane administracyjne}
\usage{
S2_mies_ad1(x, raport, mies)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{raport}{rodzaj raportu: po roku (wartość 0) lub po 2 latach (wartość 1)}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik}
}
\value{
lista
}
\description{
Funkcja licząca odsetek absolwentów o danym statusie
edukacyjno-zawodowym z uwzględnieniem bezrobocia (wskaźnik S2) w danym
miesiącu \code{mies} dla raportu dla danego roku \code{raport}. Wyróżniamy
następujące statusy:
\itemize{
\item{stat1 - Uczy się oraz nie pracuje na podstawie umowy o pracę i nie
jest bezrobotny (dla uczących się niestacjonarnie)}
\item{stat2 - Uczy się oraz poza rejestrem ZUS}
\item{stat3 - Uczy się (niestacjonarnie) oraz bezrobotny nie na stażu}
\item{stat4 - Uczy się (niestacjonarnie) oraz bezrobotny na stażu}
\item{stat5 - Uczy się oraz pracuje na podstawie umowy o pracę}
\item{stat6 - Nie uczy się oraz pracuje}
\item{stat7 - Nie uczy się oraz bezrobotny na stażu}
\item{stat8 - Nie uczy się oraz bezrobotny nie na stażu}
\item{stat9 - Nie uczy się oraz nie pracuje i nie jest bezrobotny}
\item{stat10 - Nie uczy się oraz poza rejestrem ZUS}
}
}
