% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_admin_1.R
\name{Z4_ods_prac_mies_ad1}
\alias{Z4_ods_prac_mies_ad1}
\title{Obliczanie wskaznikow dla 1. fali monitoringu - dane administracyjne}
\usage{
Z4_ods_prac_mies_ad1(x, raport = 0, od = 9, do = 12, nauka)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{raport}{rodzaj raportu: po roku (wartość 0) lub po 2 latach (wartość 1)}

\item{od}{początek okresu wyliczania wskaźnika wyrażony miesiącem}

\item{do}{koniec okresu wyliczania wskaźnika wyrażony miesiącem}

\item{nauka}{wartość TRUE/FALSE określająca czy status ma być liczony dla
absolwentów uczących się czy nie uczących się}
}
\value{
lista
}
\description{
Funkcja licząca na potrzeby szablonu raportu odsetek pracujących
w danym miesiącu w podziale na pobierających i nie pobierających nauki. W
raporcie w 2021 roku będzie to "Odsetek miesięcy, w których absolwenci
pracowali od \emph{września 2020} do \emph{grudnia 2020} roku.
}
