% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_admin_1.R
\name{Z8_formy_prac_mies_ad1}
\alias{Z8_formy_prac_mies_ad1}
\title{Obliczanie wskaznikow dla 1. fali monitoringu - dane administracyjne}
\usage{
Z8_formy_prac_mies_ad1(x, raport = 0, mies = 12, nauka)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{raport}{rodzaj raportu: po roku (wartość 0) lub po 2 latach (wartość 1)}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik}

\item{nauka}{wartość TRUE/FALSE określająca czy status ma być liczony dla
absolwentów uczących się czy nie uczących się}
}
\value{
lista
}
\description{
Funkcja licząca odsetek absolwentów wykonujących dane formy
pracy w danym miesiącu (w raportach jest to grudzień). Wskaźnik może być
liczony albo dla absolwentów pracujących i kontunuujących naukę (\code{nauka
= TRUE}) lub dla absolwentów pracujących i nie kontunuujących nauki
(\code{nauka = FALSE}). Formy pracy:
\itemize{
\item{\code{ucz_uop_niesam}  - Uczący się, zatrudnieni na podstawie umowy o
pracę, brak samozatrudnienia},
\item{\code{ucz_nieuop_sam}  - Uczący się, prowadzący działalność
gospodarczą (samozatrudnienie), brak umowy o pracę},
\item{\code{ucz_uop_sam}  - Uczący się, zatrudnieni na podstawie umowy o
pracę i prowadzący działalność gospodarczą},
\item{\code{nieucz_uop_nieinne}  - Nieuczący się, zatrudnieni na podstawie
umowy o pracę, brak innej formy},
\item{\code{nieucz_sam_nieinne}  - Nieuczący się, prowadzący działalność
gospodarczą (samozatrudnienie), brak innej formy},
\item{\code{nieucz_inne}  - Nieuczący się, zatrudnieni w innej formie (umowa
zlecenie)},
\item{\code{nieucz_uop_sam}  - Nieuczący się, zatrudnieni na podstawie umowy
o pracę i prowadzący działalność gospodarczą},
\item{\code{nieucz_uop_inne} - Nieuczący się, zatrudnieni na podstawie umowy
o pracę i zatrudnieni w innej formie},
\item{\code{nieucz_sam_inne} - Nieuczący się, działalność gospodarcza i
zatrudnienie w innej formie},
\item{\code{nieucz_uop_sam_inne}  - Nieuczący się, zatrudnienie na podstawie
umowy o pracę i działalność gospodarcza i zatrudnienie w innej formie},
}
}
