% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_admin_1.R
\name{Z9_kont_mlod_ad1}
\alias{Z9_kont_mlod_ad1}
\title{Obliczanie wskaznikow dla 1. fali monitoringu - dane administracyjne}
\usage{
Z9_kont_mlod_ad1(x, raport = 0, mies = 9)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{raport}{rodzaj raportu: po roku (wartość 0) lub po 2 latach (wartość 1)}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik}
}
\value{
lista
}
\description{
Funkcja licząca odsetek absolwentów, którzy w danym miesiącu (w
raporcie jest to wrzesień) kontynuowali zatrudnienie u pracodawcy, który
wcześniej (kiedy byli jeszcze uczniami) zatrduniał ich jako pracowników
młodocianych. Wyróżnianych jest 5 form kontynuowania pracy u pracodawcy
zatrudniającego poprzednio jako młodocianego:
\itemize{
\item{\code{nieucz_niekontuop} - Nieuczący się, nie kontynuujący pracy u
danego pracodawcy}
\item{\code{nieucz_kont_uop} - Nieuczący się, kontynuujący pracę u danego
pracodawcy na podstawie umowy o pracę}
\item{\code{nieucz_kont_inne} - Nieuczący się, kontynuujący pracę u danego
pracodawcy w formie innej niż umowa  o pracę}
\item{\code{ucz_niekontuop} - Uczący się, nie kontynuujący pracy u danego
pracodawcy na podstawie umowy o pracę}
\item{\code{ucz_kontuop} - Uczący się, kontynuujący pracę u danego
pracodawcy na podstawie umowy o pracę}
}
}
