% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki_poziom_zagregowany_admin_1.R
\name{zawody_S3_ad1}
\alias{zawody_S3_ad1}
\title{Obliczanie wskaznikow dla 1. fali monitoringu - dane administracyjne}
\usage{
zawody_S3_ad1(x, raport = 0, mies = 12)
}
\arguments{
\item{x}{ramka danych pośrednich P3}

\item{raport}{rodzaj raportu: po roku (wartość 0) lub po 2 latach (wartość 1)}

\item{mies}{miesiąc, dla którego ma być policzony wskaźnik}
}
\value{
lista
}
\description{
Funkcja licząca odsetek absolwentów o danym statusie
edukacyjno-zawodowym wśród grup absolwentów wydzielonych ze względu na zawód,
w którym się kształcili w szkole. Zwracana lista służy jako wsad do tabeli w
raporcie automatycznym. Domyślnie funkcja ma zwracać wsad do tabeli dla
grudnia 2020, ale można to zmienić poprzez zmianę argumentów funkcji. Jeśli
lista będąca wsadem tabeli generowałaby pustą tabelę, to zwracana jest
wartość \code{NULL}. W przeciwnym wypadku zwracana jest lista.
}
\details{
#' @title Obliczanie wskaznikow dla 1. fali monitoringu - dane administracyjne
#' @description Funkcja licząca odsetek absolwentów o danym statusie
#' edukacyjno-zawodowym (wskaźnik S3) w danym miesiącu \code{mies} dla raportu
#' dla danego roku \code{raport}. Wyróżniamy następujące statusy (bez KUZ i
#' KKZ):
#' \itemize{
#'  \item{stat1 - Uczy się i pracuje na podstawie umowy o pracę lub jest
#'  samozatrudniony}
#'  \item{stat2 - Uczy się i nie pracuje na podstawie umowy o pracę ani nie jest
#'  samozatrudniony}
#'  \item{stat3 - Nie uczy się i pracuje}
#'  \item{stat4 - Nie uczy się i nie pracuje}
#'  \item{stat5 - Uczy się i poza rejestrem ZUS}
#'  \item{stat6 - Nie uczy się i poza rejestrem ZUS}
#' }
#' Funkcja powinna być wykorzystana dla miesięcy od stycznia do października.
#' @param x ramka danych pośrednich P3
#' @param raport rodzaj raportu: po roku (wartość 0) lub po 2 latach (wartość 1)
#' @param mies miesiąc, dla którego ma być policzony wskaźnik
#' @return lista
#' @importFrom dplyr \%>\% filter .data summarise n_distinct
S3_mies_sty_paz_ad1 = function(x, raport, mies) {
stopifnot(is.data.frame(x),
raport \%in\% c(0, 1),
mies \%in\% c(1:12))

mies = 2020 * 12 + mies

x \%>\%
filter(.data$OKRES \%in\% mies) \%>\%
summarise(
n = n_distinct(.data$ID_ABS),
stat1 = sum(.data$NAUKA_SZK_ABS \%in\% 1 & .data$PRACA > 0, na.rm = TRUE) / n_distinct(.data$ID_ABS),
stat2 = sum(.data$NAUKA_SZK_ABS \%in\% 1 & .data$PRACA \%in\% 0, na.rm = TRUE) / n_distinct(.data$ID_ABS),
stat3 = sum(.data$NAUKA_SZK_ABS \%in\% 0 & .data$PRACA > 0, na.rm = TRUE) / n_distinct(.data$ID_ABS),
stat4 = sum(.data$NAUKA_SZK_ABS \%in\% 0 & .data$PRACA \%in\% 0, na.rm = TRUE) / n_distinct(.data$ID_ABS),
stat5 = sum(.data$NAUKA_SZK_ABS \%in\% 1 & .data$STATUS_NIEUSTALONY \%in\% 1, na.rm = TRUE) / n_distinct(.data$ID_ABS),
stat6 = sum(.data$NAUKA_SZK_ABS \%in\% 0 & .data$STATUS_NIEUSTALONY \%in\% 1, na.rm = TRUE) / n_distinct(.data$ID_ABS)
) \%>\%
as.list() \%>\%
return()
}
}
